Create Battleship.js
This is a basic text-based Battleship game. You can run this code in an HTML file in your browser. It generates a 5x5 grid, places a battleship of size 3 at a random location, and allows the player to make guesses until they sink the battleship. It keeps track of the number of guesses and displays a congratulatory message when the battleship is sunk. You can further enhance this game by adding more features and a graphical interface as needed.